integrate partial vrp api

+++ src/backend_java/src/main/java/com/ducvu/backend_java/service/DepotService.java
@@ -2,20 +2,18 @@ package com.ducvu.backend_java.service;
 
 
 import com.ducvu.backend_java.dto.request.DepotCreateRequest;
+import com.ducvu.backend_java.dto.request.DepotUpdateRequest;
 import com.ducvu.backend_java.dto.response.DepotResponse;
-import com.ducvu.backend_java.dto.response.OsmResponse;
-import com.ducvu.backend_java.dto.response.VehicleResponse;
 import com.ducvu.backend_java.model.Depot;
+import com.ducvu.backend_java.model.Vehicle;
 import com.ducvu.backend_java.repository.DepotRepository;
 import com.ducvu.backend_java.repository.VehicleRepository;
-import com.ducvu.backend_java.util.Helper;
 import com.ducvu.backend_java.util.Mapper;
 import com.ducvu.backend_java.util.Validator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -26,7 +24,6 @@ public class DepotService {
   private final VehicleRepository vehicleRepository;
   private final Validator validator;
   private final Mapper mapper;
-  private final Helper helper;
 
   public DepotResponse getDepot(String depotId) {
     Depot depot = depotRepository.findById(depotId)
@@ -39,6 +36,19 @@ public class DepotService {
         .stream()
         .map(mapper::map)
         .toList();
+  }
+
+  public DepotResponse updateDepot(String depotId, DepotUpdateRequest request) {
+    validator.validate(request);
+
+    Depot depot = depotRepository.findById(depotId)
+        .orElseThrow(() -> new RuntimeException("Depot not found"));
+
+    depot.setLatitude(request.getLatitude());
+    depot.setLongitude(request.getLongitude());
+    depot.setAddress(request.getAddress());
+
+    return mapper.map(depotRepository.save(depot));
   }
 
   public DepotResponse createDepot(DepotCreateRequest request) {
@@ -63,6 +73,13 @@ public class DepotService {
   }
 
   public void deleteDepot(String depotId) {
+    Depot depot = depotRepository.findById(depotId)
+            .orElseThrow(() -> new RuntimeException("Depot not found"));
+
+    if (!depot.getVehicles().isEmpty()) {
+      throw new RuntimeException("Cannot delete depot with assigned vehicles");
+    }
+
     depotRepository.deleteById(depotId);
   }
